+++ blend/crypto/src/lib.rs
@@ -1,5 +1,5 @@
 use blake2::{Blake2b512, digest::Digest as _};
-use lb_utils::blake_rng::{BlakeRng, RngCore as _, SeedableRng as _};
+use lb_utils::chacha_rng::{ChaCha20Rng, RngCore as _, SeedableRng as _};
 
 pub mod cipher;
 pub mod merkle;
@@ -7,7 +7,7 @@ pub mod merkle;
 pub type ZkHash = lb_groth16::Fr;
 pub type ZkHasher = lb_poseidon2::Poseidon2Bn254Hasher;
 
-/// Generates random bytes of the constant size using [`BlakeRng`].
+/// Generates random bytes of the constant size using [`ChaCha20Rng`].
 #[must_use]
 pub fn random_sized_bytes<const SIZE: usize>() -> [u8; SIZE] {
     let mut buf = [0u8; SIZE];
@@ -16,23 +16,32 @@ pub fn random_sized_bytes<const SIZE: usize>() -> [u8; SIZE] {
 }
 
 pub fn fill_random_bytes(buf: &mut [u8]) {
-    BlakeRng::from_entropy().fill_bytes(buf);
+    ChaCha20Rng::from_entropy().fill_bytes(buf);
 }
 
 /// Generates pseudo-random bytes of the constant size
-/// using [`BlakeRng`] which is seeded with a hash of the domain and key.
+/// using [`ChaCha20Rng`] which is seeded with a hash of the domain and key.
 #[must_use]
 pub fn pseudo_random_sized_bytes<const SIZE: usize>(domain: &[u8], key: &[u8]) -> [u8; SIZE] {
     let mut buf = [0u8; SIZE];
-    blake_random_bytes(&mut buf, domain, key);
+    pseudo_random_bytes(&mut buf, domain, key);
     buf
 }
 
 /// Writes pseudo-random bytes to the given buffer,
-/// using [`BlakeRng`] which is seeded with a hash of the domain and key.
-fn blake_random_bytes(buf: &mut [u8], domain: &[u8], key: &[u8]) {
-    let mut cipher = BlakeRng::from_seed(blake2b512(&[domain, key]).into());
-    cipher.fill_bytes(buf);
+/// using [`ChaCha20Rng`] which is seeded with a hash of the domain and key.
+fn pseudo_random_bytes(buf: &mut [u8], domain: &[u8], key: &[u8]) {
+    let mut rng = chacha20_rng(&blake2b512(&[domain, key]));
+    rng.fill_bytes(buf);
+}
+
+/// Builds the CSPRBG defined by the ChaCha20-Based PRNG Construction of the
+/// spec: a [`ChaCha20Rng`] keyed with the first 32 bytes of the seed.
+#[must_use]
+pub fn chacha20_rng(seed: &[u8; 64]) -> ChaCha20Rng {
+    let mut key = [0u8; 32];
+    key.copy_from_slice(&seed[..32]);
+    ChaCha20Rng::from_seed(key)
 }
 
 /// Computes the BLAKE2b-512 hash of the concatenated inputs.
