+++ blend/message/src/fixtures/mod.rs
@@ -81,12 +81,12 @@ codec_fixtures!(
             VerifiedProofOfQuota::from_bytes_unchecked([0u8; PROOF_OF_QUOTA_SIZE]),
             VerifiedProofOfSelection::from_bytes_unchecked([0u8; PROOF_OF_SELECTION_SIZE]),
         ).unwrap()], 1
-    ).unwrap() => "47a7f32151949c60050ec4454b43fcaf351a2f2383ddef2a6ab4176e269e34477821d1abaf629a228ad07b998628f4dd1e137827ca30ec8d99e90aaf9ff355af72e911fbe5eaaf7a867ca80e0a45d5a00c89a7360996aaf496503291d771adeb9caed0ca2bc20af7c31ecea182b4eb797300b68a4e5001ee438e45b402993984782478001f7336041173182189484d18804b75fb1b753c8c7cc0ae56d45c1d5b281ed36752418b833ac7e8d97bb2f78a3ac0ef9704c4f4c61ebd1c2bbfb3806dabbd2ef7b33c7778ce23a4133ac0dcf3d39c43f0562090f590506fd30e38eae7b8eb89690481bbb9a9848921d9d951b56a4ad15eec0093997cf07c04722b32edccf3bec96815f21a40d1e40e7fe5cea75d821f9763339402a92e136541b6837c7e"
+    ).unwrap() => "2b70b8a6701e1fb124ef53aeb30edbfc76f0a9553a244eccb8701b31ce152ebc25fe2a69e4862373850d41acf2677775c548c68be6994be07a892e8fa51392d2c31e68a76603f23640fe3f1773a9d4831180fe01023c4bc77e5d3061ad8c5da8b628f43e02f03186c0cecd6dea071230e86ecc8fbfb49847c2fbfc1690377de8a7b1fbbbb57a531e20c8f7a31b1403cfc99bfd0b6f1a6757cb7c6f7d86cc00b95b02236237583bf2294a8c00e2efbb312dd3b8b8dc07318b3fe42f6d720dd1f7a4884a57593a858ae7a08d8f9ee6b85dc27dea3c85d0acc515267f16baae8903abd259c07aa07536e0729a7f6d5e78d85b49c0d77127f8f11d78317832d86271da62e70be90ee52ce5d8ad3edb0cc25d63ce26614d77a33611cb1ffd51bb882c55"
 );
 
 codec_fixtures!(
     EncapsulatedBlendingHeader,
-    Self::initialize(&BlendingHeader::pseudo_random(&[1u8; 32])) => "9c05c033b59c05091fe7e3bd1bbc4abc77f69a71ba282f10e6675bb051fd4a2a5875f96f219d2a8fadfd1ed5af6476c35d6be6a485647abf2709569a69b0aa1c030ca0e01eed758d1158e36e487c1282ec22aa61edbfdd7032067707f477951f27e8f5e76ef941dde3de62a3cea5a9b6ed3e5294a836c6cb6b50eda8a09a27f57fe0e30ffcc8b93c159708ac9a230eda377da92d675c8db8e95c2f9009e1ae63e10278547404d7b335a4a749f2811570e34d6588123e1c2a614faabd0602a3fb4698466e50305eafcddebcd175c63bd560e47c883a993c87a9d9460db9ca83c56602f1eee75124c22637baa8d9fc42a62b3a34317dd4ab53e3d71442dadec7b1e8a4c258d208a3d3662a4f83674f93d6074a81e902c33d4adcbd1c995f85281200"
+    Self::initialize(&BlendingHeader::pseudo_random(&[1u8; 32])) => "33573c0931c2434ba29b86c026fdcc8cba744100c7bb4bdaf1ef55bbd5e4425a819bf64f4c12d05fd95d24731a53a6703dc4cd9dc044905f956589e89aaa882920c4546147079da7062d1e12fbe0f7754196fdce784cf5af594b49f6bbe0e6e4a630776efe360909612c2044ce97a37c07d1fa9e6bcb46034e32b7e6ca66c67d26ddb8f85f1c4fc844c233f079227766089e3ef4ae662bca0c7a79f70b38785ba2c066418939a7a9edf2080b05a1ef75377b19bf3d468678e371ae2bcb1873480115578dbf551afa1ed7a64d4e2a6c6916f0506347248ca2a1353d4a583dd14b67a76e5ea36d2edbc82de90b7f4ccb16b8e1dcac201459479c72bf75e4752465c68571fb9fa6cc8cd6f742e054bf591e93ac3b63f248187adf323273766ee52100"
 );
 
 codec_fixtures!(
