+++ blend/provers/src/crypto/test_utils.rs
@@ -156,7 +156,7 @@ impl<CorePoQGenerator> CoreLeaderAndPowProofsGenerator<CorePoQGenerator>
             Some(BlendLayerProof {
                 proof_of_quota: VerifiedProofOfQuota::from_bytes_unchecked([0; _]),
                 proof_of_selection: VerifiedProofOfSelection::from_bytes_unchecked([0; _]),
-                ephemeral_signing_key: UnsecuredEd25519Key::generate_with_blake_rng(),
+                ephemeral_signing_key: UnsecuredEd25519Key::generate_with_chacha_rng(),
             })
         })
     }
