+++ services/blend/src/edge/backends/libp2p/tests/redials.rs
@@ -30,7 +30,7 @@ async fn edge_drops_message_after_exhausting_attempts() {
         EdgeSwarmBuilder::new(Membership::new_without_local(from_ref(&Node {
             address: empty_multiaddr.clone(),
             id: random_peer_id,
-            public_key: UnsecuredEd25519Key::generate_with_blake_rng().public_key(),
+            public_key: UnsecuredEd25519Key::generate_with_chacha_rng().public_key(),
         })))
         .with_max_dial_attempts(3)
         .build();
@@ -87,7 +87,7 @@ async fn edge_redial_uses_exponential_backoff() {
         EdgeSwarmBuilder::new(Membership::new_without_local(from_ref(&Node {
             address: empty_multiaddr,
             id: random_peer_id,
-            public_key: UnsecuredEd25519Key::generate_with_blake_rng().public_key(),
+            public_key: UnsecuredEd25519Key::generate_with_chacha_rng().public_key(),
         })))
         .build();
     let message = TestEncapsulatedMessage::new(b"test-payload");
@@ -151,7 +151,7 @@ async fn stalled_send_does_not_block_command_processing() {
     } = EdgeSwarmBuilder::new(Membership::new_without_local(from_ref(&Node {
         address,
         id: peer_id,
-        public_key: UnsecuredEd25519Key::generate_with_blake_rng().public_key(),
+        public_key: UnsecuredEd25519Key::generate_with_chacha_rng().public_key(),
     })))
     .build();
