+++ services/blend/src/edge/handlers.rs
@@ -11,7 +11,7 @@ use lb_blend::{
     },
 };
 use lb_chain_service::Epoch;
-use lb_utils::blake_rng::BlakeRng;
+use lb_utils::chacha_rng::ChaCha20Rng;
 use overwatch::overwatch::OverwatchHandle;
 use rand::SeedableRng as _;
 
@@ -98,7 +98,7 @@ where
             settings.backend,
             overwatch_handle,
             membership,
-            BlakeRng::from_entropy(),
+            ChaCha20Rng::from_entropy(),
             settings.non_ephemeral_signing_key,
         );
         Self {
