+++ services/blend/src/test_utils/libp2p.rs
@@ -56,12 +56,12 @@ impl DerefMut for TestEncapsulatedMessage {
 }
 
 fn generate_valid_inputs() -> Vec<EncapsulationInput> {
-    repeat_with(UnsecuredEd25519Key::generate_with_blake_rng)
+    repeat_with(UnsecuredEd25519Key::generate_with_chacha_rng)
         .take(NUM_BLEND_LAYERS as usize)
         .map(|recipient_signing_key| {
             let recipient_signing_pubkey = recipient_signing_key.public_key();
             EncapsulationInput::try_new(
-                UnsecuredEd25519Key::generate_with_blake_rng(),
+                UnsecuredEd25519Key::generate_with_chacha_rng(),
                 &recipient_signing_pubkey,
                 VerifiedProofOfQuota::from_bytes_unchecked([0; _]),
                 VerifiedProofOfSelection::from_bytes_unchecked([0; _]),
