+++ utils/Cargo.toml
@@ -14,8 +14,6 @@ workspace = true
 
 [dependencies]
 async-trait    = { workspace = true }
-blake2         = { optional = true, workspace = true }
-cipher         = { workspace = true }
 const-hex      = { features = ["alloc"], workspace = true }
 futures        = { optional = true, workspace = true }
 humantime      = { optional = true, workspace = true }
@@ -23,6 +21,7 @@ lb-log-targets = { workspace = true }
 multiaddr      = { workspace = true }
 overwatch      = { workspace = true }
 rand           = { features = ["getrandom"], workspace = true }
+rand_chacha    = { optional = true, workspace = true }
 serde          = { features = ["alloc", "derive"], workspace = true }
 serde_ignored  = { workspace = true }
 serde_with     = { optional = true, workspace = true }
@@ -33,7 +32,7 @@ tokio          = { optional = true, workspace = true }
 tracing        = { workspace = true }
 
 [features]
-rng              = ["dep:blake2"]
+rng              = ["dep:rand_chacha"]
 serde            = ["const-hex/alloc", "serde/alloc"]
 time             = ["dep:humantime", "dep:serde_with", "dep:time"]
 tokio            = ["dep:futures", "dep:tokio"]
@@ -41,8 +40,8 @@ tokio-task-names = ["tokio", "tokio/rt", "tokio/tracing"]
 
 [dev-dependencies]
 bincode     = { workspace = true }
+blake2      = { workspace = true }
 nistrs      = { workspace = true }
-rand_chacha = { workspace = true }
 serde_json  = { features = ["alloc"], workspace = true }
 serde_with  = { features = ["macros"], workspace = true }
 tempfile    = { workspace = true }
